Consolidated sync utilities
