+++ app/clients/local/controllers/util/hash.js
@@ -1,0 +1,9 @@
+var fs = require("fs-extra");
+var crypto = require("crypto");
+
+module.exports = function hash(path) {
+  return crypto
+    .createHash("md5")
+    .update(fs.readFileSync(path))
+    .digest("hex");
+};
